+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -201,7 +201,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg6ncw9qc060y2bv1k2s1yhm"),
+  _id: z.string().default("conv_01kg7a8phngafxxq182xat3x52"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),