+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-30T05:18:02.987Z",
+    "2026-01-30T11:22:44.854Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {