+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -94,7 +94,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-30T05:18:06.030Z",
+    "2026-01-30T11:22:48.265Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {